Split out Associations into its own section for scaling later
