+++ mathics/builtin/associations/__init__.py
@@ -1,0 +1,7 @@
+"""
+Associations
+
+An Association maps keys to values and is similar to a dictionary in Python.
+It is often sparse in that its key space is much larger than the number of \
+actual keys found in the collection.
+"""
